% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copmare_full.R
\name{compare_full}
\alias{compare_full}
\title{A vectorized version of \code{waldo::compare()}}
\usage{
compare_full(
  new = enlist_rds(private_path("pacta_core")),
  old = enlist_rds(private_path("web_tool")),
  ...
)
}
\arguments{
\item{new, old}{A list of dataframes.}

\item{...}{Passed to \code{waldos::compare()}}
}
\value{
Called for its side effect. Prints differences.
}
\description{
Defaults to comparing private results between the web tool and this package.
}
\examples{
new <- list(x = data.frame(a = 1, b = 1), x = data.frame(a = 1, b = 9))
old <- list(x = data.frame(a = 1, b = 1), x = data.frame(a = 1, b = 1))
compare_full(new, old)
compare_full(new, new)
}
\seealso{
Other developer oriented: 
\code{\link{create_pacta}()},
\code{\link{run_web_tool}()}
}
\concept{developer oriented}
