% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_env.R
\name{create_env_from_renviron}
\alias{create_env_from_renviron}
\title{Create a pacta environment file from pacta variables set in .Renviron}
\usage{
create_env_from_renviron(path = path_temp(".env"))
}
\arguments{
\item{path}{String. Path to the environment file you want to create.}
}
\description{
Create a pacta environment file from pacta variables set in .Renviron
}
\examples{
env <- create_env_from_renviron()
readLines(env)
}
\keyword{internal}
