% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_working_dir}
\alias{create_working_dir}
\title{Create a working_dir/ directory with the folders structure that pacta expects}
\usage{
create_working_dir(dir = tempdir())
}
\arguments{
\item{dir}{String. Path to where to create a pacta project.}
}
\description{
Create a working_dir/ directory with the folders structure that pacta expects
}
\examples{
dir <- tempdir()
create_working_dir(dir)
fs::dir_tree(fs::path(dir, "working_dir"))
}
\keyword{internal}
