% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissable_slow_tests.R
\name{enable_slow_tests}
\alias{enable_slow_tests}
\title{Enable/dissable slow tests}
\usage{
enable_slow_tests()
}
\description{
Enable/dissable slow tests
}
\examples{
if (interactive()) {
  enable_slow_tests()
  dissable_slow_tests()
}
}
\keyword{internal}
