% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_pacta.R
\name{example_input_paths}
\alias{example_input_paths}
\title{Path to an example porftolio and parameter file}
\usage{
example_input_paths()
}
\value{
String. A vector of paths to example portfolio and parameter files.
}
\description{
Path to an example porftolio and parameter file
}
\examples{
example_input_paths()
}
\keyword{helpers}
