% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_pacta.R
\name{get_env}
\alias{get_env}
\title{Get the value of an environment variable, or fail with an informative error}
\usage{
get_env(var)
}
\arguments{
\item{var}{String. The name of an environment variable, e.g. "PACTA_DATA".}
}
\value{
String. The value of the environment value (or an informative error).
}
\description{
Get the value of an environment variable, or fail with an informative error
}
\examples{
get_env("SHELL")
try(get_env("BAD"))
}
\seealso{
Other helpers: 
\code{\link{example_input_paths}()},
\code{\link{wd_path}()}
}
\concept{helpers}
