% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_pacta.R
\name{getenv_data}
\alias{getenv_data}
\title{Get the environment variable "PACTA_DATA" or fail gracefully}
\usage{
getenv_data()
}
\description{
Get the environment variable "PACTA_DATA" or fail gracefully
}
\examples{
getenv_data()

Sys.setenv(PACTA_DATA = "")
try(getenv_data())
}
\keyword{internal}
