% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_pacta.R
\name{local_pacta}
\alias{local_pacta}
\alias{create_pacta}
\title{Create a pacta project}
\usage{
local_pacta(dir = tempdir(), data = getenv_data(), envir = parent.frame())

create_pacta(dir = tempdir(), data = getenv_data())
}
\arguments{
\item{dir}{String. Path where to create pacta files and directories.}

\item{data}{String. Path to the private pacta-data/ directory.}

\item{env}{Must be passed on to \code{withr::defer()}.}
}
\description{
\itemize{
\item \code{create_pacta()} creates a pacta project.
\item \code{local_pacta()} creates a temporary pacta project.
}
}
\details{
See https://testthat.r-lib.org/articles/test-fixtures.html#local-helpers.
}
\examples{
# This local context is usually a call to test_that()
dir <- fs::path(tempdir(), "pacta")
fs::dir_exists(dir)

local({
  fs::dir_exists(dir)
  local_pacta(dir)
  fs::dir_exists(dir)
})

# Gone
fs::dir_exists(dir)

# Persistent
create_pacta(dir)
fs::dir_exists(dir)

# Explore
fs::dir_tree(dir, all = TRUE)
writeLines(readLines(fs::path(dir, ".env")))
}
\keyword{internal}
