% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta_envvar.R
\name{pacta_envvar}
\alias{pacta_envvar}
\title{Get enviroment variables succinctly and in a specific order}
\usage{
pacta_envvar(...)
}
\description{
... Strings. Text to match the name of the environment variables of a pacta
project.
}
\examples{
pacta_envvar()
pacta_envvar("out", "in")
pacta_envvar("in", "out")
}
\keyword{internal}
