% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta_ls.R
\name{pacta_permissions}
\alias{pacta_permissions}
\title{Explore pacta paths}
\usage{
pacta_permissions(env = NULL, ...)
}
\arguments{
\item{env}{String. Path to environment file. \code{NULL} defaults to ".env".}

\item{...}{Passed on to \code{fs::dir_ls()}.}
}
\description{
Explore pacta paths
}
\examples{
dir <- local_pacta()
withr::local_dir(dir)

pacta_ls(env)
pacta_info()

pacta_permissions(recurse = TRUE)
}
