% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_web_tool.R
\name{run_web_tool}
\alias{run_web_tool}
\title{Run the web_tool for regression tests}
\usage{
run_web_tool(
  results = path_temp(wd_path()),
  source = get_env("PACTA_ANALYSIS"),
  data = get_env("PACTA_DATA"),
  x = 1:3
)
}
\arguments{
\item{results}{String. Path where to save the results.}

\item{source}{String. Path to the PACTA_analysis repo locally.}

\item{data}{String. Path to the}

\item{x}{Number. Vector giving the web tool script to run, e.g. \code{1} or \code{1:3}.}
}
\value{
Called for its side effect. Returns the first argument invisibly.
}
\description{
Run the web_tool for regression tests
}
\examples{
\donttest{
  run_web_tool()
}
}
\seealso{
Other developer oriented: 
\code{\link{compare_full}()},
\code{\link{create_pacta}()}
}
\concept{developer oriented}
