% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fast.R
\name{test_fast}
\alias{test_fast}
\title{Skip slow tests}
\usage{
test_fast(...)
}
\arguments{
\item{...}{Arguments passed to \code{devtools::test()} and \code{devtools::check()}.}
}
\value{
Called for their side effect.
}
\description{
Skip slow tests
}
\examples{
\dontrun{
  test_fa
}
}
\seealso{
Other developer oriented: 
\code{\link{compare_full}()},
\code{\link{create_pacta}()},
\code{\link{run_web_tool}()},
\code{\link{update_pacta_legacy}()}
}
\concept{developer oriented}
