% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pacta_legacy.R
\name{update_pacta_legacy}
\alias{update_pacta_legacy}
\title{Update inst/extdata/context/pacta_legacy.R with scripts from PACTA_analysis}
\usage{
update_pacta_legacy(scripts = NULL)
}
\value{
Called for its side effect of writting a file. Returns the first
argument invisibly.
}
\description{
Update inst/extdata/context/pacta_legacy.R with scripts from PACTA_analysis
}
\examples{
\dontrun{
  # Online
  update_pacta_legacy()

  # Local
  scripts <- path("~", "PACTA_analysis", paste0("web_tool_script_", 1:2, ".R"))
  update_pacta_legacy(scripts)
}
}
\seealso{
Other developer oriented: 
\code{\link{compare_full}()},
\code{\link{create_pacta}()},
\code{\link{run_web_tool}()},
\code{\link{test_fast}()}
}
\concept{developer oriented}
