% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_web_tool.R
\name{wd_path}
\alias{wd_path}
\title{Path based on working_dir, to hide this low level implementation detail}
\usage{
wd_path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[fs:path]{fs::path}}
  \describe{
    \item{\code{ext}}{An optional extension to append to the generated path.}
  }}
}
\value{
String. A path.
}
\description{
This function avoids littering the code base with specific paths we want to
move away from.
}
\examples{
wd_path()
}
\seealso{
Other helpers: 
\code{\link{example_input_paths}()},
\code{\link{get_env}()}
}
\concept{helpers}
