% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_working_dir}
\alias{create_working_dir}
\title{Create the structure of pacta's working_dir/}
\usage{
create_working_dir(dir = tempdir())
}
\description{
Create the structure of pacta's working_dir/
}
\examples{
dir <- create_working_dir()
fs::dir_tree(fs::path(dir, "working_dir"))
}
