% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_env.R
\name{local_env}
\alias{local_env}
\alias{create_env}
\title{Help create an ephemeral environment file}
\usage{
local_env(path = fs::path_temp(".env"), ..., envir = parent.frame())

create_env(
  path = fs::path_temp(".env"),
  output = fs::path_temp("output"),
  input = fs::path_temp("input"),
  data = getenv_data()
)
}
\arguments{
\item{path}{String. Path to the environment file you want to create.}

\item{...}{Passed on to \code{create_env()}.}

\item{env}{Must be passed on to \code{withr::dever()}.}
}
\description{
See https://testthat.r-lib.org/articles/test-fixtures.html#local-helpers.
}
\examples{
# The local context is usually a call to test_that()
env <- tempfile("env_")
local({
  env <- local_env(
    path = env,
    input = "a/b",
    output = "c/d",
    data = "e/f/pacta-data"
  )
  fs::file_exists(env)
  readLines(env)
})

# Gone
fs::file_exists(env)

# Persists
create_env(
  path = env,
  input = "a/b",
  output = "c/d",
  data = "e/f/pacta-data"
)
fs::file_exists(env)
readLines(env)
}
\keyword{inernal}
\keyword{internal}
