% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_pacta.R
\name{local_pacta}
\alias{local_pacta}
\alias{create_pacta}
\title{Create a pacta project}
\usage{
local_pacta(
  dir = tempdir(),
  input_paths = example_input_paths(),
  data = getenv_data(),
  envir = parent.frame()
)

create_pacta(
  dir = tempdir(),
  input_paths = example_input_paths(),
  data = getenv_data()
)
}
\arguments{
\item{dir}{String. Path where to create pacta files and directories.}

\item{data}{String. Path to the private pacta-data/ directory.}

\item{env}{Must be passed on to \code{withr::defer()}.}
}
\description{
\itemize{
\item \code{create_pacta()} creates a pacta project.
\item \code{local_pacta()} creates a temporary pacta project.
}
}
\details{
See https://testthat.r-lib.org/articles/test-fixtures.html#local-helpers.
}
\examples{
dir <- path(tempdir(), "pacta")
withr::local_envvar(c(PACTA_DATA = path(dir, "pacta-data")))

# This local context is usually a call to test_that()
local({
  local_pacta(dir)
  dir_tree(dir, all = TRUE)
})

dir_exists(dir)
}
\keyword{internal}
