% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta_ls.R
\name{pacta_ls}
\alias{pacta_ls}
\alias{pacta_permissions}
\alias{pacta_info}
\title{Explore pacta paths}
\usage{
pacta_ls(env = NULL, ...)

pacta_permissions(env = NULL, ...)

pacta_info(env = NULL, ...)
}
\arguments{
\item{env}{String. Path to environment file. \code{NULL} defaults to ".env".}

\item{...}{Passed on to \code{fs::dir_ls()}.}
}
\description{
Explore pacta paths
}
\examples{
# Setup
dir <- tempdir("pacta")

withr::with_dir(dir, {
fs::dir_create("pacta-data")
local_pacta(dir, data = fs::path_temp("pacta-data"))

# Inspect
readLines(".env")

pacta_ls()
pacta_info()

pacta_permissions(recurse = TRUE)
})
}
\keyword{internal}
