% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta_core.R
\name{run_pacta_core}
\alias{run_pacta_core}
\title{Run PACTA on Docker}
\usage{
run_pacta_core()
}
\value{
Called for its side effects.
}
\description{
Run PACTA on Docker
}
\examples{
if (interactive()) {
  run_pacta_core()
}
}
