% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta_core.R
\name{run_pacta_core}
\alias{run_pacta_core}
\title{Run pacta core}
\usage{
run_pacta_core(env)
}
\description{
Run pacta core
}
\examples{
if (interactive()) {
  dir <- withr::local_tempdir("pacta")
  withr::local_dir(dir)
  local_pacta(dir)

  fs::dir_tree(all = TRUE)
  readLines(".env")

  run_pacta_core(".env")
}
}
