% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta_core.R
\name{run_pacta_core}
\alias{run_pacta_core}
\title{Run the core steps of the PACTA methodology}
\usage{
run_pacta_core(env = ".env")
}
\arguments{
\item{env}{String. Path to an environment file with pacta variables.}
}
\value{
Called for its side effect. Returns the first argument inbisibly.
}
\description{
Run the core steps of the PACTA methodology
}
\examples{
if (interactive()) {
  dir <- fs::path_abs("../pacta")
  create_pacta(dir)

  env <- fs::path(dir, ".env")
  readLines(env)

  run_pacta_core(env)

  fs::dir_tree(dir)
}
}
