% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta_core.R
\name{run_pacta_core}
\alias{run_pacta_core}
\title{Run the core steps of the PACTA methodology}
\usage{
run_pacta_core(env = ".env")
}
\arguments{
\item{env}{String. Path to an environment file with pacta variables.}
}
\value{
Called for its side effect. Returns the first argument invisibly.
}
\description{
Run the core steps of the PACTA methodology
}
\examples{
if (interactive()) {
  local({
    # Create a toy, temporary pacta project
    pacta <- fs::path_home("pacta_tmp")
    local_pacta(pacta)

    # Work from the pacta project for convenicenc
    withr::local_dir(pacta)

    # The ".env" file specifies the setup for input/ output/ and private data
    readLines(".env")

    # Before
    fs::dir_tree()
    run_pacta_core()
    # After
    fs::dir_tree()
  })
}
}
